refactor(test): P5-1 follow-up — extract Tier-4 OOS / perf-hints
Continue test_widget_bridge.cpp split. Extracts the 138-line Tier-4
catalog-hygiene cluster (pulp #1434 / Tier-4) into a sibling TU.

Two sub-clusters bundled together since both pin no-op + fallback
behavior for OOS properties:
  1. OOS 3D / generated-content / scroll-snap pin — properties Pulp
     deliberately doesn't paint (3D transforms, ::before/::after,
     scroll-snap-type / -align). Pins no-op + non-crash contract.
  2. Perf hints + interaction misc pin (Bundle-C) — will-change,
     contain, content-visibility, touch-action secondaries,
     pointer-events refinements. Pins acceptance + fallback.

* test_widget_bridge.cpp: 6,849 → 6,711 lines (-138)
* New target pulp-test-widget-bridge-tier4-oos: 175 lines, 2
  TEST_CASEs, 6 assertions.

This is commit 2 of 2 in a bundled-cuts PR (RN style-prop primitives
shipped in the preceding commit; single version bump comes next).
